Fix bug when top-left cell starts adjacent to a same-color cell
If one of the neighbors of the top-left cell was picked to be the same
color as the top-left cell, they wouldn't be connected in the first
color update since connecting cells by color was done _after_ updating
all cells in the same set as the top-left cell.

The fix here was to just swap the order of things in the
ChangeTopLeftColor update: update connections by color before updating
the colors.  This did not break victory checks because that just checks
that all cells are the same color, but if that ever changes to check
that all cells belong to the same set, this will need to be revisited.